Fix double delete for file_case usage..
This fixes (Encapsulate searchpath:: FILE*'s in new class
file_case), the very first commit of this series.
At that time i falsely thought that the ownership is not overtaken
but that recursion occurs, so that the caller is still responsible
for the file_case* pointer, which is plain wrong, since
a file_iterator object is still created and that always overtakes
ownership.