[mdocmx] mdocmx.{awk,7}: actually understand mdoc(7); improve awk(1)..
Until now we only knew about those mdoc(7) commands which the new
.Mx command will create anchors for; also we didn't care for
punctuation and where incapable to support "command
continuations", so that the following would have failed in _all_
aspects:

  .It Ic argone , argtwo , Bt bla

We should now be able to skip the commas, recognize that "argtwo"
is a KEY that belongs to the MACro "Ic", understand that "Bt"
cancels continuation of "Ic" in that "bla" is not a KEY of "Ic",
etc.

Also do recognize .Sh and .Ss commands and create anchors for
them, so that the mdoc(7) builtin command .Sx will also be able to
create references to them.

The awk(1) was pretty weird, too, take advantage of associative
mappings and actually lookup commands via hashtable instead of
walking arrays.

Finally: rename mdocmx.1 to mdocmx.7 and adhere to the new stuff.