[mdocmx] mdocmx(7): drop ".Mx *DIGITS" syntax..
because it clashes (GNU) mdoc(7) parsing.  In

  .Bl -tag -width ".It Va _utoprin_"
  .Mx *3
  .It Va agent-shell-lookup-USER@HOST , agent-shell-lookup-HOST , \
    agent-shell-lookup

The ".It Va _utoprin_" that will specify the width is actually
parsed at a time that interferes with the desired .Mx usage.
And whereas that only applies to the first .It and can be worked
around with some effort it is not really sane to provide something
that requires _such_ special care.