[mdocmx] mdocmx.sh: allow $AWK overwrites; properly deal with SIGPIPE..
Ooops, we yet didn't allow an environmental $AWK setting!

And then: awk(1) doesn't support signal handlers, which means
that, when we're part of a pipe which the user terminates, we are
not capable to deal with the broken pipe case that our END{}
handler will generate when we had to perform any preprocessing,
and that in turn would result in a dangling temporary file!

Thus the only sane option seems to be to always create the
temporary file, wether we need it or not, not to exec(1) awk(1)
but keep on running the shell in order to remove the temporary
after awk(1) has finished, whichever way.

Since from within awk(1) we would have needed to system()
invocations in case any preprocessing had to be performed it is
pretty likely that doing it this way is a (rather massive)
improvement on the resource side, too.