[mdocmx] mdocmx.sh: don't strip leading dot but for 1st macro..
When it is the turn to parse a macro while checking a line
for the possibility to pop the stack, we yet stripped a leading
dot from the macro, regardless of were we where on the line.
This (of course) doesn't work for constructs like the following
from the (S-roff) mdoc(7) manual:

  .Mx
  .Mx
  .It Ic .Es , .En

The second .En was not considered for popping because as 'En' it
was believed to cause macro evaluation and thus end the .Ic block.