[mdocmx] mdocmx.tmac: cut down performance impact..
One of the to-be-expected acceptance problems of mdocmx(7) is the
performance impact that is induced by resolving all the
references.

Whereas that isn't even noticeable for manuals with a small number
of references the times of course explode exponentially with
a growing number of them, because we need to store them in
"singly-linked lists", rather than being able to use the natural
choice, hashmaps – or a similar structure.

However, groff(1) allows for one more single identification of
a string, its `.length'.  So add an additional subdivision layer
of our singly-linked lists that forms sublists of strings of equal
size; of course this won't help a roff(7) manual that solely
documents two letter strings, but it does help in practice, even
though the benefit can be expected to be much larger for other
real-life manuals, even in the absence of a `.hash' roff(7)
command.
For completeness sake extend mdocmx.tmac already today for
troff(1)s which do support .hash, which essentially should drop
the list search counts near hit-and-run.

Before:
  ?0[]$ time (mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >/dev/null)
      0m1.25s real     0m1.59s user     0m0.06s system
  ?0[]$ time (mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >/dev/null)
      0m1.29s real     0m1.60s user     0m0.06s system

  ?0[]$ time (export MDOCMX_ENABLE=1; mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >MAN.before)
      0m7.36s real     0m7.73s user     0m0.08s system
  ?0[]$ time (export MDOCMX_ENABLE=1; mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >MAN.before)
      0m7.36s real     0m7.74s user     0m0.08s system

After:
  ?0[]$ time (mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >/dev/null)
      0m1.19s real     0m1.59s user     0m0.06s system
  ?0[]$ time (mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >/dev/null)
      0m1.20s real     0m1.59s user     0m0.05s system

  ?0[]$ time (export MDOCMX_ENABLE=1; mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >MAN.old)
      0m2.59s real     0m2.98s user     0m0.07s system
  ?0[]$ time (export MDOCMX_ENABLE=1; mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >MAN.old)
      0m2.56s real     0m2.97s user     0m0.07s system

After, with an experimental .hash troff(1) builtin:
  ?0[]$ time (export MDOCMX_ENABLE=1; mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >MAN.new)
      0m2.24s real     0m2.62s user     0m0.07s system
  ?0[]$ time (export MDOCMX_ENABLE=1; mdocmx.sh nail.1|
  groff -Tutf8 -mdoc -dmx-toc-force=1 -dmx-debug=1 >MAN.new)
      0m2.19s real     0m2.61s user     0m0.07s system

For comparison, here is plain mdocml(1):
  ?0[]$ time (mdocmx.sh nail.1|mandoc > /dev/null)
      0m0.26s real     0m0.19s user     0m0.07s system

So yes, the impact is still ~100% for the *extremely* mdocmx(7)
stressing S-nail manual with its 4500 rendered lines at the time
of testing.  But i don't say ouch, because the user experience is
simply not comparable!  *And* i presume that mdocml(1) wouldn't
cross the 0m0.30s even if it had mdocmx(7) support, and then we're
completely on the rockin' side of things.