[mdocmx] mdocmx.sh: with -t or -T, define mx-toc-name..
Just in case the mdocmx.tmac macros do not exist we need to
ensure we have the "TABLE OF CONTENTS" string!
Problem is (again) compatibility with mdocml.bsd.lv (mandoc(1)),
which does not even support the simplemost "if !d NAME" conditionals,
so we actually either loose user-defined strings or mandoc(1) compat,
of which i currently rate the latter as (much) worse.