Sync-to-go: tmac..
Practically in sync.
I did not take/integrate the Bernd Warken commit series that
turned copyrights etc. into macros, nor his one which uses strings
etc. for apostrophes etc. (i saw errors in that one, too).
On the long run i want to convert most manuals into mdoc(7), and
thus that is counterproductive.
Another reason for the pretty large resulting diff(1) is that
i already started to break up lines more nicely etc.
But we are in effect in sync.

Note that the following list is partial.

- tmac/an-ext.tmac
  . Removed useless version information.
- tmac/doc.tmac
  . Blindly took [9125552] ([mdoc] Fix Savannah bug #44733.,
    2015-04-11, Ingo Schwarze).
- man/tmac-man.7.in
  . Ensure compatibility mode works.
  . Took [8b8a9de] (groff_man.man: Formatting., 2014-03-27, Bjarni
    Ingi Gislason).
  . Did not take [f64b964] (groff/tmac/groff_man.man: Minor typos,
    formatting, reordering., 2014-01-05) except for some typos and
    formatting stuff.  Then i sorted alphabetically and did some
    more tweaks (macros preceeded by dot etc.), so we are there.
- man/tmac-me.7.in
  . Manual inclusion of [25fcdee] (Changed to use TBL and
    eliminate low-level troff hackery., 2012-07-28, Eric S.
    Raymond) and [0896bca] (groff_me.man: Formatting., Based on
    a patch from Bjarni Ingi Gislason <bjarniig@rhi.hi.is>.,
    2014-03-27).