Sync-to-go: man..
This changeset touches quite some files that were already updated
by other [sync.2015-05-20] changesets: this is because i renamed
the l_roff_XY to l_roff-XY, and (dit)?roff.7.in to l_ditroff.7.in
and l_roff-history.7.in, and updated all references accordingly.

Practically in sync.
For now we however miss requests like .class and .lsm, so that
those must be subtracted from this statement.

I did not take/integrate the Bernd Warken commit series that
turned copyrights etc. into macros, nor his one which uses strings
etc. for apostrophes etc. (i saw errors in that one, too).
On the long run i want to convert most manuals into mdoc(7), and
thus that is counterproductive.
Another reason for the pretty large resulting diff(1) is that
i already started to break up lines more nicely etc.
The following is a partial list:

- l_roff.7.in
  . [6054a4e] (groff.7: add documentation for the Unicode escapes
    \[u....] and \[u.....], 2014-08-05, Bernd Warken) i did
    differently.
  . I haven't taken the new sections on underlining etc.
  . I have taken the text of [e85eb55], (* man/groff.man: Document
    missing number registers (`$$', `.b', `.F', `.j', `.k', `.L',
    `.O', `.P', `.R', `opmaxx', `opmaxy', `opminx', `opminy').,
    2009-02-10).
- l_roff-diff.7.in
  . Taken [9af742c] (* man/groff_diff.man: Document \n[.P].,
    2010-06-05).
- l_roff-out.5.in
  . Took (some formatting parts of) [21eee56] (groff_out.man:
    Formatting, whitespace., 2014-03-27, Bjarni Ingi Gislason).
- l_roff-tmac.5.in
  . Took (some formatting parts of) [03149c3] (groff_tmac.man:
    Formatting, whitespace., 2014-03-27, Bjarni Ingi Gislason).