Sync-to-go: src/pre-grn..
Practically in sync.

TODO - I've not taken [4b0a4fc], but force a panic if that many
TODO   file arguments happen to happen (50!).
TODO   (From the adaption there is already a TODO note to switch
TODO   the code to our getopt and then i think we maybe have
TODO   a vector to use for this, too.)

Despite this upper-limit adjustment this changeset only fixes
false whitespace introduced in adaption.