[mdocmx.1.fixes] mdocmx.sh: colon : is a special built-in, ++..
therefore a redirect error will cause a hard shell exit and we
need to use a separated subshell (), {;} is not enough.

Rise number of temporary file attempts to three even in verbose
mode.

And we do not need to rm(1) the temporary file in the signal
catching traps since those exit, which in turn will invoke the
normal EXIT trap.  (At least in non-buggy shells.)