[mdocmx.1.fixes] mdocmx.tmac: FIX -ix (Apologies: Ingo Schwarze)..
Oh.  It didn't work.  -ix did not use the (valid) data of the
preprocessor, but instead always generated anchors by itself,
resulting in dangling references before the anchor and mess in
general.  So instead do not process -ix requests if the document
is not preprocessed, and simply dump an anchor otherwise.

While here, fix problems caused by automatic string instantiation
upon use, so do not use ".if d STRING" if there is a chance that
STRING may (have) be(en) auto-instantiated.