[hastybuild] Introduce su/: Steffen's Utilities..
Since i will _never_ find the time to port my C++ library (to
a C library) but need stuff *now* and have needed some stuff for
the S-nail MUA i maintain already in the past, start providing
a simple framework that can be interchanged in between the
projects the future.

Start now with the basic header(s).