[hastybuild] NO! Rethink SU (SU umbrella, II)..
The initial commit of this was on 2018-01-16 22:15:28 +0100.

  After fooling around for weeks i decided to make a cut and rethink
  what the SU foundation is about to do.
  It shall be more than thought initially.
  I want to have a lot more comfort.

  And i want that the SU tools can optionally be "hardened" in that
  by setting some configuration bits memory allocation failures do
  not cause panic()s but return NIL/ERR_NOMEM all up the call chain.
  And ditto, i want that all the SU tools that matter (like strings
  and collections) can handle datatype overflow in the very same
  way: panic() by default, but by setting a bit ERR_OVERFLOW errors
  shall be reported all up the call chain.
  This latter change means that existing code needs to be rewritten
  to reflect return type changes.

  In addition i never have done it like that, placing almost each
  function in its own file.  No.  Instead group them logically as
  far as possible, and so that fewest possible object dependencies
  result in order to show good behaviour when linking statically.
  (Even though one of my goals is to end up with only libraries,
  which are driven by small command line handling executables.)

  Also, introduce all the nice little things that my C++ library
  had, with code-in.h/code-ou.h pairs which define shorthands, etc.

  This delays all even longer, but i think the positive effects will
  be worth it and spring into existence sooner rather than later.

In the meantime i have not worked on groff in all this time
(2022-08-31), and also the mailer had huge holes.  (Corona virus
and more.  Getting older, heh.  Did a lot here and there, went
back to Linux and learned about it, etc.)

For the mailer however i started porting my old C++ SF stuff to
some extend for real, a tad differently however, as we have to
integrate in C library runtimes etc etc.  And of course, only as
much as is needed to drive the mailer forward.  The entire I/O,
network, format / brackex / date / you name it, all that is
missing.  The memory cache ;>  But regardless, we went a bit, and
it will get more (if i live long enough).

This 2nd umbrella commit will be used to embed that SU stuff once
i take it over, until then that 2018 code will remain.
(Most of the non-purely-SU things can remain as such even, it
seems!)