[mdocmx.2] tmac/: do apply mdocmx to m?doc* stuff _for_us_..
Funnily i have never applied it in this repo, likely to be able to sync
better, or whatever i thought.  All this time i had MBOXes with less and
groff merge requests (2014 and 2015, respectively, until 2022-08-31),
but these i dropped yesterday, so, and in hindsight that groff has
a new maintainer who implemented OSC 8 (with a \X^tty: link^ one!),
and that less now at least tolerates OSC 8, but with my _new_ merge
request "could do it", and we rewrote the thing for OSC 8 not BS
sequences (which would not have been able to do it, i think col(1)
was the reason why, i have forgotten details; it was a hack!), simply
merge the thing in from my GNU groff.git [mdocmx] topic branch with
timestamps 2015-01-07 14:43:18 +0100 - 2015-02-05 15:56:36 +0100,
with a .Fc fix on 2015-05-01 16:08:53 +0200.  (I never updated that
one for all the changes made in this repo here, i simply copied over
to the 1.22.3 installation.)